con patients por cada user

+++ lib/services/consultation_service.dart
@@ -1,14 +1,23 @@
 import 'package:med_copilot_1/models/consultation.dart';
 import 'dart:convert';
 import 'package:http/http.dart' as http;
+import 'package:med_copilot_1/services/auth_service.dart';
 
 class ConsultationService {
+  final AuthService _authService = AuthService();
   final String baseUrl = 'http://localhost:5000/consultations';
+  String? token;
+  int? user;
 
   // Obtener todas las consultas
-  Future<List<Consultation>> fetchConsultations() async {
-    final response = await http.get(Uri.parse(baseUrl));
+  Future<List<Consultation>> fetchConsultations() async {    
+    token = await _authService.getToken();
+    user = await _authService.getUser();
+    if (token == null) {
+      throw Exception('No se encontró un token. Inicia sesión nuevamente.');
+    }
 
+    final response = await http.get(Uri.parse('$baseUrl/user/$user'));
     if (response.statusCode == 200) {
       List<dynamic> data = json.decode(response.body);
       return data.map((json) => Consultation.fromJson(json)).toList();
@@ -18,9 +27,13 @@ class ConsultationService {
   }
 
   // Obtener todas las consultas por pacientes
-  Future<List<Consultation>> fetchConsultationsByPatient(int patientId) async {
-    final response = await http.get(Uri.parse('$baseUrl/patient/$patientId'));
+  Future<List<Consultation>> fetchConsultationsByPatient(int patientId) async {    
+    final token = await _authService.getToken();
+    if (token == null) {
+      throw Exception('No se encontró un token. Inicia sesión nuevamente.');
+    }
 
+    final response = await http.get(Uri.parse('$baseUrl/patient/$patientId'));
     if (response.statusCode == 200) {
       List<dynamic> data = json.decode(response.body);
       return data.map((json) => Consultation.fromJson(json)).toList();
@@ -30,7 +43,12 @@ class ConsultationService {
   }
 
   // Agregar una nueva consulta
-  Future<void> createConsultation(Consultation consultation) async {
+  Future<void> createConsultation(Consultation consultation) async {    
+    final token = await _authService.getToken();
+    if (token == null) {
+      throw Exception('No se encontró un token. Inicia sesión nuevamente.');
+    }
+
     final response = await http.post(
       Uri.parse(baseUrl),
       headers: {'Content-Type': 'application/json'},
@@ -43,7 +61,12 @@ class ConsultationService {
   }
 
   // Actualizar una consulta existente
-  Future<void> updateConsultation(Consultation consultation) async {
+  Future<void> updateConsultation(Consultation consultation) async {    
+    final token = await _authService.getToken();
+    if (token == null) {
+      throw Exception('No se encontró un token. Inicia sesión nuevamente.');
+    }
+
     final response = await http.put(
       Uri.parse('$baseUrl/${consultation.id}'),
       headers: {'Content-Type': 'application/json'},
@@ -60,7 +83,12 @@ class ConsultationService {
   }
 
   // Eliminar una consulta
-  Future<void> deleteConsultation(int id) async {
+  Future<void> deleteConsultation(int id) async {    
+    final token = await _authService.getToken();
+    if (token == null) {
+      throw Exception('No se encontró un token. Inicia sesión nuevamente.');
+    }
+    
     final response = await http.delete(Uri.parse('$baseUrl/$id'));
 
     if (response.statusCode == 404) {
